+++ meituan/Others/3tdParts/LoopView/ImageScrollCell.m
@@ -13,7 +13,7 @@
 -(id)initWithStyle:(UITableViewCellStyle)style reuseIdentifier:(NSString *)reuseIdentifier{
     self = [super initWithStyle:style reuseIdentifier:reuseIdentifier];
     if (self) {
-        NSLog(@"imageArr:%ld",self.imageArr.count);
+        NSLog(@"imageArr:%ld",(unsigned long)self.imageArr.count);
         self.imageScrollView = [[ImageScrollView alloc] initWithFrame:CGRectMake(0, 0, screen_width, 180) ImageArray:self.imageArr];
         [self.contentView addSubview:self.imageScrollView];
     }
@@ -23,7 +23,7 @@
 -(id)initWithStyle:(UITableViewCellStyle)style reuseIdentifier:(NSString *)reuseIdentifier frame:(CGRect)frame{
     self = [super initWithStyle:style reuseIdentifier:reuseIdentifier];
     if (self) {
-        NSLog(@"imageArr:%ld",self.imageArr.count);
+        NSLog(@"imageArr:%ld",(unsigned long)self.imageArr.count);
         self.imageScrollView = [[ImageScrollView alloc] initWithFrame:CGRectMake(0, 0, frame.size.width, frame.size.height) ImageArray:self.imageArr];
         [self.contentView addSubview:self.imageScrollView];
     }
